+++ History.md
@@ -6,6 +6,7 @@
 * BREAKING: Migration functions for past versions are no longer available
 * UI Helpers now use `console.debug` for debugging messages
 * `RolesCollection` and `RoleAssignmentsCollection` can now be exported in addition to being accessed via `Meteor.roles` and `Meteor.roleAssignment`
+* Support for Meteor 3
 
 ## v3.6.3